Distribute excess measurements 'diagonally'
When simulating reception of a number of letters which is not an integer
multiple of the number of cells in the rectangle, previous code
incorrectly distributed the 'excess' in raster order.  The correct
behaviour is to step along each axis simultaneously, wrapping when
reaching the edge.  I.e., the main movement is 'diagonally'.  Implement
this, and update test.